Moved CopyAndReleaseImageDataCommand and CopyAndReleaseBufferDataCommand into their own headers/source files
